// Features/reports/Types/IIncentiveForReport.ts
export interface IIncentiveForReport {
    identifier: string;
    idIncentive: number;
    description: string;
    chargeDate: string;
    dateExecuted: string;
    payDate: string;
    isExecuted: string;
    amount: number;
    tax: number;
    isPaid: string;
    idEmployee: number;
    idConcept: number;
    idPayrollPay: number;
    conceptCode: string;
    concept: string;
    payrollName: string;
    payrollNumber: number;
    payrollPayDate: string;
    idPosition: number;
    position: string;
    idDepartment: number;
    department: string;
    employee: string;
    idCompany: number;
}